Summarizing the content of xplain XML files

xplain.overview summarizes the content of an xplain XML file.

xplain.overview(xml, show.text=FALSE, preserve.seq=FALSE)

Arguments

xml
Path to the xplain XML file. Can be either a local path or an URL.

show.text
Indicates if the full interpretation/explanation texts shall be included in the summary (optional). Default: FALSE.

preserve.seq
Indicates if the overview results for the interpretation/explanation texts shall be shown in the same sequence as they appear in the XML file (optional). If FALSE, the results are sorted, e.g. by package, function, language and complexity level. Default: FALSE.

Value

A data frame summarizing the XML file. Each row corresponds to a <title> or <text> element in the xplain XML file.
The column structure is as follows:

Package: The package to which the explained function belongs.

Function: The explained function.

Type: Indicates whether the element is a <title> or a <text> element.

Language: Language of the element (also considering inheritance from higher-level XML elements).

Level: Complexity level of the element (also considering inheritance from higher-level XML elements).

Result object: Element of the explained function's return object to which the <title> or <text> element relates (if any).

Iteration: Type of iteration (if any). Value of the foreach attribute of the <text> element.

Has R code: Indicates if the <text> element includes R code.

Uses return obj.: Indicates if the <text> element refers to the explained function's return object.

Text: The text of the respective <text> or <title> element (including R code). This column is only included if show.text=TRUE.

To learn more about the structure of xplain XML files, go to the xplain help page or consult the web tutorial.

See also

More material on the internet

Web tutorial on how to work with xplain: http://www.zuckarelli.de/xplain/index.html
xplain cheat sheet: http://www.zuckarelli.de/xplain/xplain_cheatsheet.pdf

Examples

xplain.overview("http://www.zuckarelli.de/xplain/example_lm.xml")

#>   Package Function  Type Language Level Result object Iteration Has R code
#> 1   stats       lm title  default   all                              FALSE
#> 2   stats       lm  text  default   <=1                               TRUE
#> 3   stats       lm  text  default   <=1                               TRUE
#>   Uses return obj.
#> 1            FALSE
#> 2             TRUE
#> 3             TRUE